requirements.txt
dj-database-url==0.4.2
Django==1.11
django-braces==1.11.0
django-debug-toolbar==1.6
django-model-utils==3.0.0
django-registration-redux==1.6
django-star-ratings==0.5.5
django-taggit==0.22.1
gunicorn==19.7.1
pkg-resources==0.0.0
psycopg2==2.7.1
python-decouple==3.0
pytz==2017.2
sqlparse==0.2.3
whitenoise==3.3.0
